refactor(core): split candidate-store into focused modules
Keep the public candidate-store surface stable while splitting the 942-line
implementation into axis tokens, types, geometry helpers, eager indexes, and
a lazy construction shell. Characterization tests stay green; add a facade
export surface assertion so internal helpers cannot leak from the barrel.
